coal: Don't include hpp/fcl/config.hpp if collision is not enable
Split forward declaration in a standalone header to avoid code
duplication in sdf/geometry.hxx.
